New Feature Added to MyBatis Plugin: Annotation SQL Reload Functionality (#558)
* Fix the issue in MyBatis-Spring mode where a NullPointerException (NPE) occurs during reload, and also differentiate between MyBatis and MyBatis-Spring modes

* Update MyBatisTransformers.java

Fix the issue where using the insertAt method might fail due to changes in the processPropertyPlaceHolders method body. Add a list variable at the beginning of the method, include the factory in the list after factory initialization, and finally execute the removal logic at the end of the method.

Why not execute the removal right after initialization? I tried that, but it fails during reload.

* Enhance MyBatis Plugin: Code Optimization and New Features

1.Refactor MyBatisTransformers code to improve readability and maintainability.
3.Feature Added: Enable SQL reload when using annotation-based SQL.
2.Feature Added: Enable reload when there are changes in entity classes, such as result entity or parameter entity classes.

* Fix: Issue with spring-mybatis reload not working with annotated SQL

---------

Co-authored-by: hejiajin <Homejim5>