Fix and optimize MyBatis Plus plugin tests. (#562)
* Fix the issue in MyBatis-Spring mode where a NullPointerException (NPE) occurs during reload, and also differentiate between MyBatis and MyBatis-Spring modes

* Update MyBatisTransformers.java

Fix the issue where using the insertAt method might fail due to changes in the processPropertyPlaceHolders method body. Add a list variable at the beginning of the method, include the factory in the list after factory initialization, and finally execute the removal logic at the end of the method.

Why not execute the removal right after initialization? I tried that, but it fails during reload.

* 1.Fix MybatisSpringBootTest:Fix MybatisSpringBootTest to prevent exception logs when running unit tests.
2.Enable MyBatisPluginAnnoTest unit test
3.Add --add-opens=java.base/sun.nio.ch=ALL-UNNAMED to fix the error "Unable to make field protected final xxx" when running unit tests with jbr17.

---------

Co-authored-by: hejiajin <Homejim5>