Upgrade Mybatis Version Compatibility to 3.5.16 and Refactor Common Test Methods into BaseTest (#565)
* Fix the issue in MyBatis-Spring mode where a NullPointerException (NPE) occurs during reload, and also differentiate between MyBatis and MyBatis-Spring modes

* Update MyBatisTransformers.java

Fix the issue where using the insertAt method might fail due to changes in the processPropertyPlaceHolders method body. Add a list variable at the beginning of the method, include the factory in the list after factory initialization, and finally execute the removal logic at the end of the method.

Why not execute the removal right after initialization? I tried that, but it fails during reload.

* 1.Added a timeout value to the WaitHelper.waitForCommand method in MyBatisPluginAnnoTestEnhanced version compatibility by upgrading the supported version from 3.5.9 to 3.5.16. The main change occurred in version 3.5.12, where constructor parameters were modified.
                                                                                      2.Refactored common test methods into a base class BaseTest to facilitate future use.

---------

Co-authored-by: hejiajin <Homejim5>