feat(mybatis-plus): Add MyBatisPlus plugin to HA (#567)
* Fix the issue in MyBatis-Spring mode where a NullPointerException (NPE) occurs during reload, and also differentiate between MyBatis and MyBatis-Spring modes

* Update MyBatisTransformers.java

Fix the issue where using the insertAt method might fail due to changes in the processPropertyPlaceHolders method body. Add a list variable at the beginning of the method, include the factory in the list after factory initialization, and finally execute the removal logic at the end of the method.

Why not execute the removal right after initialization? I tried that, but it fails during reload.

* add mybatis plus plugin

* support versions between 3.4.0 and 3.5.7

* change package path

---------

Co-authored-by: hejiajin <Homejim5>