Fix Mybatis plugin: Remove potential code errors. (#607)
* Fix the issue in MyBatis-Spring mode where a NullPointerException (NPE) occurs during reload, and also differentiate between MyBatis and MyBatis-Spring modes

* Update MyBatisTransformers.java

Fix the issue where using the insertAt method might fail due to changes in the processPropertyPlaceHolders method body. Add a list variable at the beginning of the method, include the factory in the list after factory initialization, and finally execute the removal logic at the end of the method.

Why not execute the removal right after initialization? I tried that, but it fails during reload.

* Remove the logging code; in lower versions of spring-mybatis, there is no getMapperInterface method, which causes errors.

---------

Co-authored-by: hejiajin <Homejim5>